create an internal package and move MorphiaUtils there.
